+++ runtime/lib/simd128.cc
@@ -456,6 +456,15 @@ DEFINE_NATIVE_ENTRY(Int32x4_and, 0, 2) {
   return Int32x4::New(_x, _y, _z, _w);
 }
 
+DEFINE_NATIVE_ENTRY(Int32x4_not, 0, 1) {
+  GET_NON_NULL_NATIVE_ARGUMENT(Int32x4, self, arguments->NativeArgAt(0));
+  int32_t _x = ~self.x();
+  int32_t _y = ~self.y();
+  int32_t _z = ~self.z();
+  int32_t _w = ~self.w();
+  return Int32x4::New(_x, _y, _z, _w);
+}
+
 DEFINE_NATIVE_ENTRY(Int32x4_xor, 0, 2) {
   GET_NON_NULL_NATIVE_ARGUMENT(Int32x4, self, arguments->NativeArgAt(0));
   GET_NON_NULL_NATIVE_ARGUMENT(Int32x4, other, arguments->NativeArgAt(1));