+++ runtime/vm/bootstrap_natives.h
@@ -228,6 +228,7 @@ namespace dart {
   V(Int32x4_or, 2)                                                             \
   V(Int32x4_and, 2)                                                            \
   V(Int32x4_xor, 2)                                                            \
+  V(Int32x4_not, 1)                                                            \
   V(Int32x4_add, 2)                                                            \
   V(Int32x4_sub, 2)                                                            \
   V(Int32x4_getX, 1)                                                           \