+++ sdk/lib/_internal/wasm/common/simd.dart
@@ -636,6 +636,7 @@ final class I32x4 extends WasmTypedDataBase implements Int32x4 {
       I32x4.fromV128(_bits & (other as I32x4)._bits);
   Int32x4 operator ^(Int32x4 other) =>
       I32x4.fromV128(_bits ^ (other as I32x4)._bits);
+  Int32x4 operator ~() => I32x4.fromV128(~_bits);
   Int32x4 operator +(Int32x4 other) => I32x4.fromV128(
     (WasmI32x4(_bits) + WasmI32x4((other as I32x4)._bits)).value,
   );